+++ src/Plainion.GraphViz.Modules.CodeInspection/Packaging/Services/PackageAnalyzer.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Threading;
 using Plainion.GraphViz.Modules.CodeInspection.Core;
 using Plainion.GraphViz.Modules.CodeInspection.Packaging.Spec;
@@ -14,7 +12,7 @@ namespace Plainion.GraphViz.Modules.CodeInspection.Packaging.Services
     class PackageAnalyzer
     {
         private static string[] Colors = { "LightBlue", "LightGreen", "LightCoral", "Brown", "DarkTurquoise", "MediumAquamarine", "Orange", 
-                                             "LawnGreen","DarkKhaki","BurlyWood","SteelBlue","Goldenrod", "Tomato","Crimson","CadetBlue" };
+                                           "LawnGreen","DarkKhaki","BurlyWood","SteelBlue","Goldenrod", "Tomato","Crimson","CadetBlue" };
 
         private SystemPackaging myConfig;
         private CancellationToken myCancellationToken;